switch  between recruiter and student accounts feature implemented
When you have both a student and recruiter account under the same email, you will be able to login once and switch between those 2 accounts without logging out and logging in with  a different account. The application will detect upon login if you have 2 accounts and it will log you in on a dual user profile where you will have an option to switch between accounts. If you only have 1 account, the app will detect and log you into a different user profile for that specific role you logged in as, without features of switching to another account.